services/vision-ai/requirements.txt
# Core dependencies
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
werkzeug==3.0.1
numpy<2.0.0
pillow==10.1.0

# Vision and AI dependencies
torch==2.1.0
torchvision==0.16.0
transformers==4.36.0
sentence-transformers==2.2.2

# For multimodal models
openai==1.3.0

# gRPC dependencies
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1

# Image processing
opencv-python==4.8.1.78
